requirements.txt
pypdf
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
huggingface_hub
flask
faiss-cpu
sentence-transformers
langchain-classic
langchain_classic